tui_client/requirements.txt
imgui
glfw
PyOpenGL
pyarrow
